requirements_versions_py312_windows.txt
setuptools==69.5.1  # temp fix for compatibility with some old packages
GitPython==3.1.32
Pillow==10.4.0
accelerate==1.9.0
blendmodes==2024.1.1
clean-fid==0.1.35
diskcache==5.6.3
einops==0.4.1
facexlib==0.3.0
fastapi==0.94.0
gradio==3.41.2
httpcore==0.15
inflection==0.5.1
jsonmerge==1.8.0
kornia==0.6.7
lark==1.1.2
numpy==1.26.4
omegaconf==2.2.3
open-clip-torch==3.1.0
piexif==1.1.3
protobuf==7.34.1
psutil==5.9.5
pytorch_lightning==1.9.4
resize-right==0.0.2
safetensors>=0.6.2
scikit-image>=0.22.0
spandrel==0.3.4
spandrel-extra-arches==0.1.1
tensorflow==2.20.0
timm>=1.0.17
tomesd==0.1.3
torch
torchdiffeq==0.2.3
torchsde==0.2.6
transformers==5.4.0
httpx==0.24.1
pillow-avif-plugin==1.4.3
peft==0.15.2

# ADetailer (extensions-builtin): was install.py; unified here
ultralytics>=8.3.75
rich>=13.0.0
huggingface_hub>=0.20.0

# WD14-tagger (extensions-builtin): was install.py; unified here
# Only packages NOT already pinned or implied elsewhere are listed here.
deepdanbooru
jsonschema
opencv_contrib_python
